drivers: misc: Fix coverity issues
Fix 2 coverity issues.
CID 10165044
CID 10165045

Bug 3952896

Reviewed-on: https://git-master.nvidia.com/r/c/linux-nv-oot/+/2850604
GVS: Gerrit_Virtual_Submit <buildbot_gerritrpt@nvidia.com>

+++ drivers/misc/nvscic2c-pcie/common.h
@@ -1,5 +1,5 @@
 /* SPDX-License-Identifier: GPL-2.0-only */
-/* Copyright (c) 2022, NVIDIA CORPORATION & AFFILIATES. All rights reserved. */
+/* Copyright (c) 2022-2023, NVIDIA CORPORATION & AFFILIATES. All rights reserved. */
 
 #ifndef __COMMON_H__
 #define __COMMON_H__
@@ -200,6 +200,6 @@ enum peer_cpu_t {
 /* Returns aperture offset of syncpoint on SHIM_BASE. */
 static inline u64 get_syncpt_shim_offset(u32 id)
 {
-	return (SHIM_BASE + (id * SP_SIZE));
+	return (SHIM_BASE + ((u64)id * SP_SIZE));
 }
 #endif //__COMMON_H__
